cleaned up unused code and renamed folders and files
