--- src/components/cart-item 2.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-const CartItem = () => <div></div>;
-
-export default CartItem;
