+++ src/components/payment-form.tsx
@@ -1,8 +1,7 @@
-import React, { useState, useEffect } from "react";
-import { Box, Form, Grid, RadioButtonGroup } from "grommet";
+import React, { useState } from "react";
+import { Box, Form, RadioButtonGroup } from "grommet";
 
-import FormFieldLabel from "./formFieldLabel";
-import { type } from "os";
+import FormFieldLabel from "./form-field-fabel";
 
 interface IProps {}
